+++ agent-control/tests/on_host/scenarios/empty_config.rs
@@ -1,4 +1,4 @@
-use crate::on_host::tools::base_paths::TempBasePaths;
+use crate::common::base_paths::TempBasePaths;
 use crate::on_host::tools::config::create_remote_config;
 use crate::{
     common::{
