+++ agent-control/tests/on_host/tools.rs
@@ -1,4 +1,3 @@
-pub mod base_paths;
 pub mod config;
 pub mod custom_agent_type;
 pub mod fake_binary;
