--- agent-control/tests/on_host/tools/base_paths.rs
@@ -1,1 +1,0 @@
-pub use crate::common::base_paths::TempBasePaths;
